# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: PermsCraft REST API
  description: |
    Embedded REST API for the PermsCraft Minecraft permission plugin.

    ## Authentication
    Every request must include a valid API key in the `Authorization` header:
    ```
    Authorization: Bearer <api-key>
    ```
    Keys are managed in-game or from console:
    ```
    /pc apikey create <label> [scopes...]
    /pc apikey list
    /pc apikey revoke <label>
    ```

    ## Scopes
    | Scope    | Description |
    |----------|-------------|
    | `read`   | GET endpoints for groups, users, server info |
    | `write`  | POST/PUT/DELETE for groups and users |
    | `log`    | Access to audit log endpoints |
    | `sync`   | Trigger cache flush and sync operations |
    | `backup` | Export, import, and restore backups |
    | `admin`  | All scopes combined |

    ## Security
    > ⚠️ The API transmits keys over plain HTTP by default.
    > It is strongly recommended to use a TLS reverse proxy (nginx/Caddy)
    > and bind to `127.0.0.1` in `config.yml`.
    > See the [TLS setup guide](https://github.com/your-repo/wiki/REST-API-TLS).

    ## Rate Limiting
    Default: 120 requests per minute per IP. Configurable in `config.yml`.
  version: 1.0.0
  contact:
    name: PermsCraft
    url: https://github.com/your-repo
servers:
  - url: http://localhost:4567/api/v1
    description: Local server (default)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Group 'admin' not found."
    Group:
      type: object
      properties:
        name:
          type: string
          example: admin
        weight:
          type: integer
          example: 100
        prefix:
          type: string
          example: "&c[Admin] "
        suffix:
          type: string
          example: ""
        parents:
          type: array
          items:
            type: string
          example: ["default", "vip"]
        permissions:
          type: array
          items:
            type: string
          example: ["essentials.fly", "-essentials.ban"]
        meta:
          type: object
          additionalProperties:
            type: string
    User:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        username:
          type: string
          example: Steve
        groups:
          type: array
          items:
            type: string
          example: ["admin", "vip"]
        permissions:
          type: array
          items:
            type: string
        prefix:
          type: string
        suffix:
          type: string
        meta:
          type: object
          additionalProperties:
            type: string
    TimedPermission:
      type: object
      properties:
        permission:
          type: string
          example: "essentials.fly"
        expires:
          type: integer
          format: int64
          description: Unix timestamp (seconds)
          example: 1735689600
...
paths:
...
  /groups:
    get:
      summary: List all groups
      tags: [Groups]
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  count:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a group
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: vip
                weight:
                  type: integer
                  default: 0
                  example: 10
      responses:
        "201":
          description: Group created
        "400":
          description: Invalid input or group already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{uuid}:
    get:
      summary: Get user by UUID
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
  /users/name/{name}:
    get:
      summary: Look up user by username
      tags: [Users]
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
  /users/{uuid}/permissions/check:
    get:
      summary: Check a single permission for a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: node
          required: true
          schema:
            type: string
          example: essentials.fly
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                type: object
                properties:
                  permission:
                    type: string
                  result:
                    type: boolean
  /users/{uuid}/permissions/resolved:
    get:
      summary: Get full resolved permission map for a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Full permission map
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
  /users/{uuid}/permissions:
    post:
      summary: Add a permission to a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permission]
              properties:
                permission:
                  type: string
      responses:
        "200":
          description: Permission added
  /users/{uuid}/permissions/{perm}:
    delete:
      summary: Remove a permission from a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: perm
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Permission removed
  /users/{uuid}/groups:
    post:
      summary: Add user to a group
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [group]
              properties:
                group:
                  type: string
      responses:
        "200":
          description: User added to group
  /users/{uuid}/groups/{group}:
    delete:
      summary: Remove user from a group
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: group
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User removed from group
  /users/{uuid}/timed-permissions:
    get:
      summary: List timed permissions for a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of timed permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimedPermission'

    post:
      summary: Add a timed permission to a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permission, duration]
              properties:
                permission:
                  type: string
                  example: "essentials.fly"
                duration:
                  type: string
                  description: "Duration string: 1d, 12h, 30m, 7d12h"
                  example: "7d"
      responses:
        "200":
          description: Timed permission added
  /users/{uuid}/timed-permissions/{perm}:
    delete:
      summary: Revoke a timed permission
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: perm
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timed permission revoked
  /users/{uuid}/prefix:
    put:
      summary: Set user prefix
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prefix:
                  type: string
      responses:
        "200":
          description: Prefix set
  /users/{uuid}/suffix:
    put:
      summary: Set user suffix
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                suffix:
                  type: string
      responses:
        "200":
          description: Suffix set
  /users/{uuid}/meta:
    put:
      summary: Set a meta key on a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                value:
                  type: string
      responses:
        "200":
          description: Meta set
  /users/{uuid}/meta/{key}:
    delete:
      summary: Delete a meta key from a user
      tags: [Users]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Meta key deleted

  # ── Logs ────────────────────────────────────────────────────────────────────
...
  /sync/user/{uuid}:
    post:
      summary: Sync a specific user across servers
      tags: [Sync]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User sync triggered
...
  /cache/flush/user/{uuid}:
    post:
      summary: Flush cache for a specific user
      tags: [Sync]
      parameters:
        - in: path
          name: uuid
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User cache flushed
...